requirements.txt
boto3==1.4.7
botocore==1.7.5
click==6.7
cssmin==0.2.0
docutils==0.14
Flask==0.12.2
Flask-Assets==0.12
Flask-Cors==3.0.3
futures==3.1.1
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
jsmin==2.2.2
MarkupSafe==1.0
olefile==0.44
Pillow==4.2.1
python-dateutil==2.6.1
s3transfer==0.1.11
simplejson==3.11.1
six==1.10.0
webassets==0.12.1
Werkzeug==0.12.2
